requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
python-dotenv==1.0.0
Pillow==11.2.1
